refactor: extract mind upgrade flow into lib/mind/upgrade.ts
Moves updateTemplateBranch, mergeTemplateBranch, mergeUpgradeAndRestart,
and configureGitIdentity out of web/api/minds.ts into a new
lib/mind/upgrade.ts, exposing runUpgrade/continueUpgrade/abortUpgrade/
upgradeDiff/upgradeInProgress. The /:name/upgrade route now only parses
the request, validates the template, dispatches to the lib, and maps
outcomes back to the existing response shapes/status codes. No behavior
change.
