+++ packages/daemon/src/web/api/minds.ts
@@ -38,7 +38,6 @@ import {
   recordNotice,
 } from "../../lib/chat/system-events.js";
 import { getSpiritName } from "../../lib/config/setup.js";
-import { readSystemsConfig } from "../../lib/config/systems-config.js";
 import { getMindManager, MindStartupError } from "../../lib/daemon/mind-manager.js";
 // Lifecycle functions from mind-service.ts
 import {
@@ -88,7 +87,6 @@ import {
   readRegistry,
   removeMind,
   setMindStage,
-  setMindTemplate,
   setMindTemplateHash,
   stateDir,
   validateMindName,
@@ -96,6 +94,15 @@ import {
 import { evaluateSeedChecklist } from "../../lib/mind/seed-readiness.js";
 import { isTemplateStale } from "../../lib/mind/template-staleness.js";
 import { applyThinkingLevel, deriveThinkingLevel } from "../../lib/mind/thinking-config.js";
+import {
+  abortUpgrade,
+  configureGitIdentity,
+  continueUpgrade,
+  runUpgrade,
+  TEMPLATE_BRANCH,
+  upgradeDiff,
+  upgradeInProgress,
+} from "../../lib/mind/upgrade.js";
 import { cleanupVariant } from "../../lib/mind/variant-cleanup.js";
 import {
   findUnresolvedHomeFiles,
@@ -117,12 +124,7 @@ import {
   getImagegenJob,
   waitForImagegenJob,
 } from "../../lib/services/imagegen-jobs.js";
-import {
-  getStandardSkillsWithExtensions,
-  installSkill,
-  migrateSkillsToTemplate,
-  SEED_SKILLS,
-} from "../../lib/skills.js";
+import { getStandardSkillsWithExtensions, installSkill, SEED_SKILLS } from "../../lib/skills.js";
 import { convertSession } from "../../lib/template/convert-session.js";
 import {
   findOpenClawSession,
@@ -132,7 +134,6 @@ import {
 } from "../../lib/template/import-utils.js";
 import {
   applyInitFiles,
-  applyTemplateHomeFiles,
   composeTemplate,
   copyTemplateToDir,
   findTemplatesRoot,
@@ -292,19 +293,6 @@ export function toPublicMind(
   };
 }
 
-const TEMPLATE_BRANCH = "volute/template";
-
-/** Configure per-repo git identity for a mind: name = mind name, email = [mind].[system]@volute.systems. */
-async function configureGitIdentity(
-  mindName: string,
-  opts: { cwd: string; mindName?: string; env?: NodeJS.ProcessEnv },
-) {
-  const systemsConfig = readSystemsConfig();
-  const system = systemsConfig?.system ?? "local";
-  await gitExec(["config", "user.name", mindName], opts);
-  await gitExec(["config", "user.email", `${mindName}.${system}@volute.systems`], opts);
-}
-
 /**
  * Create the volute/template tracking branch and main branch with shared history.
  * Enables clean 3-way merges on the first `volute mind upgrade`.
@@ -330,231 +318,6 @@ async function initTemplateBranch(
   await gitExec(["checkout", "-b", "main"], opts);
   await gitExec(["add", "-A"], opts);
   await gitExec(["commit", "-m", "initial commit"], opts);
-}
-
-/**
- * Update the volute/template orphan branch with the latest template files.
- * Uses a temporary worktree to avoid touching the main working directory.
- */
-async function updateTemplateBranch(projectRoot: string, template: string, mindName: string) {
-  const tempWorktree = resolve(projectRoot, ".variants", "_template_update");
-
-  let branchExists = false;
-  try {
-    await gitExec(["rev-parse", "--verify", TEMPLATE_BRANCH], { cwd: projectRoot });
-    branchExists = true;
-  } catch {
-    // branch doesn't exist
-  }
-
-  // Clean up any existing temp worktree
-  try {
-    await gitExec(["worktree", "remove", "--force", tempWorktree], { cwd: projectRoot });
-  } catch {
-    // doesn't exist
-  }
-  if (existsSync(tempWorktree)) {
-    rmSync(tempWorktree, { recursive: true, force: true });
-  }
-
-  const templatesRoot = findTemplatesRoot();
-  const { composedDir, manifest } = composeTemplate(templatesRoot, template);
-
-  try {
-    if (branchExists) {
-      await gitExec(["worktree", "add", tempWorktree, TEMPLATE_BRANCH], {
-        cwd: projectRoot,
-      });
-    } else {
-      await gitExec(["worktree", "add", "--detach", tempWorktree], { cwd: projectRoot });
-      await gitExec(["checkout", "--orphan", TEMPLATE_BRANCH], { cwd: tempWorktree });
-      await gitExec(["rm", "-rf", "--cached", "."], { cwd: tempWorktree });
-      await gitExec(["clean", "-fd"], { cwd: tempWorktree });
-    }
-
-    if (branchExists) {
-      await gitExec(["rm", "-rf", "."], { cwd: tempWorktree }).catch(() => {});
-    }
-
-    copyTemplateToDir(composedDir, tempWorktree, mindName, manifest);
-
-    const initDir = resolve(tempWorktree, ".init");
-    if (existsSync(initDir)) {
-      rmSync(initDir, { recursive: true, force: true });
-    }
-
-    // Remove home files except VOLUTE.md — template branch should only track infrastructure
-    const homeDir = resolve(tempWorktree, "home");
-    if (existsSync(homeDir)) {
-      for (const entry of readdirSync(homeDir)) {
-        if (entry !== "VOLUTE.md") {
-          rmSync(resolve(homeDir, entry), { recursive: true, force: true });
-        }
-      }
-    }
-
-    await gitExec(["add", "-A"], { cwd: tempWorktree });
-
-    try {
-      await gitExec(["diff", "--cached", "--quiet"], { cwd: tempWorktree });
-    } catch {
-      await gitExec(["commit", "-m", "template update"], { cwd: tempWorktree });
-    }
-  } finally {
-    try {
-      await gitExec(["worktree", "remove", "--force", tempWorktree], { cwd: projectRoot });
-    } catch {
-      // best effort cleanup
-    }
-    if (existsSync(tempWorktree)) {
-      rmSync(tempWorktree, { recursive: true, force: true });
-    }
-    rmSync(composedDir, { recursive: true, force: true });
-  }
-}
-
-/**
- * Merge the template branch into the current worktree.
- * Returns true if there are merge conflicts.
- */
-async function mergeTemplateBranch(worktreeDir: string): Promise<boolean> {
-  try {
-    await gitExec(
-      ["merge", TEMPLATE_BRANCH, "--allow-unrelated-histories", "-m", "merge template update"],
-      { cwd: worktreeDir },
-    );
-    return false;
-  } catch (e: unknown) {
-    try {
-      const status = await gitExec(["status", "--porcelain"], { cwd: worktreeDir });
-      const hasConflictMarkers = status
-        .split("\n")
-        .some((line) => line.startsWith("UU") || line.startsWith("AA"));
-      if (hasConflictMarkers) return true;
-    } catch {
-      // fall through to rethrow
-    }
-    throw e;
-  }
-}
-
-/**
- * Merge the upgrade branch back into main, clean up, install deps, and restart.
- * Returns { ok, warning? } on success, throws on merge failure.
- */
-async function mergeUpgradeAndRestart(
-  mindName: string,
-  dir: string,
-  worktreeDir: string,
-  upgradeVariantName: string,
-  upgradeBranch: string,
-  template: string,
-  oldTemplate: string,
-): Promise<{ ok: true; warning?: string }> {
-  const templateChanged = template !== oldTemplate;
-  // Auto-commit any uncommitted changes in main worktree
-  const mainStatus = (await gitExec(["status", "--porcelain"], { cwd: dir })).trim();
-  if (mainStatus) {
-    await gitExec(["add", "-A"], { cwd: dir });
-    await gitExec(["commit", "-m", "Auto-commit before upgrade merge"], { cwd: dir });
-  }
-
-  const preMergeHead = (await gitExec(["rev-parse", "HEAD"], { cwd: dir })).trim();
-  await gitExec(["merge", upgradeBranch], { cwd: dir });
-
-  // Merge succeeded — everything below is best-effort cleanup/restart
-  try {
-    await cleanupVariant(upgradeVariantName, mindName, dir, worktreeDir);
-  } catch (err) {
-    log.warn(`failed to clean up upgrade worktree for ${mindName}`, log.errorData(err));
-  }
-  try {
-    await gitExec(["branch", "-D", upgradeBranch], { cwd: dir });
-  } catch {
-    // branch may already be deleted by cleanupVariant
-  }
-
-  // On an actual template switch, swap the template-owned home/ files (mechanics
-  // doc, .claude/settings.json, config.json) which the merge never touches. This
-  // must succeed *before* the DB template field is advanced: that field drives
-  // credential injection at spawn (mind-manager), so it has to stay consistent
-  // with the on-disk config. On failure, leave the field at oldTemplate and
-  // surface the failure rather than reporting a clean success.
-  let switchWarning: string | undefined;
-  if (templateChanged) {
-    try {
-      applyTemplateHomeFiles(resolve(dir, "home"), template);
-      // Move installed skills into the new template's skills dir and regenerate
-      // their shims, so they aren't stranded (invisible + shims pointing at the
-      // old path) after the switch.
-      const migratedSkills = migrateSkillsToTemplate(dir, oldTemplate, template);
-      await gitExec(["add", "home/"], { cwd: dir });
-      try {
-        await gitExec(["diff", "--cached", "--quiet"], { cwd: dir });
-      } catch {
-        await gitExec(["commit", "-m", `swap template-owned home files for ${template}`], {
-          cwd: dir,
-        });
-      }
-      await chownMindDir(dir, mindName);
-      const skillNote =
-        migratedSkills.length > 0
-          ? ` Migrated skills to the ${template} skills dir: ${migratedSkills.join(", ")}.`
-          : "";
-      switchWarning = `Switched ${oldTemplate}→${template}: config reset to ${template} defaults, mechanics doc replaced, conversation starts fresh (sessions aren't portable across runtimes).${skillNote}`;
-    } catch (err) {
-      log.warn(`failed to swap template home files for ${mindName}`, log.errorData(err));
-      return {
-        ok: true,
-        warning: `Upgrade merged but template switch ${oldTemplate}→${template} failed: ${err instanceof Error ? err.message : String(err)}. The mind is still registered as ${oldTemplate}; re-run the switch or fix home/ manually.`,
-      };
-    }
-  }
-
-  // Persist the template field only after any switch swap succeeded, so the DB
-  // stays consistent with the on-disk template files.
-  try {
-    await setMindTemplateHash(mindName, computeTemplateHash(template));
-    await setMindTemplate(mindName, template);
-  } catch (err) {
-    log.warn(`failed to update template for ${mindName}`, log.errorData(err));
-  }
-
-  // Skip npm install when the merge didn't touch dependencies — even a no-op
-  // install writes enough to freeze slow storage for a minute or more.
-  if (await npmInstallNeeded(dir, preMergeHead)) {
-    try {
-      await npmInstallAsMind(dir, mindName);
-    } catch (err) {
-      log.warn(`npm install failed after upgrade merge for ${mindName}`, log.errorData(err));
-      return {
-        ok: true,
-        warning: `Upgrade merged but npm install failed: ${err instanceof Error ? err.message : String(err)}. You may need to run npm install manually.`,
-      };
-    }
-  } else {
-    log.info(`skipping npm install for ${mindName} — dependencies unchanged by upgrade`);
-  }
-
-  // Restart mind with upgrade context
-  const manager = getMindManager();
-  try {
-    if (manager.isRunning(mindName)) {
-      await manager.stopMind(mindName);
-    }
-    manager.setPendingContext(mindName, { type: "upgraded" });
-    // Generous health budget: right after an npm install the disk cache is
-    // cold and I/O may still be saturated, so a tsx cold start can exceed the
-    // default 30s — timing out here kills the child and leaves the mind down.
-    await manager.startMind(mindName, { healthTimeoutMs: 120_000 });
-  } catch (e) {
-    return {
-      ok: true,
-      warning: `Upgrade merged but mind restart failed: ${e instanceof Error ? e.message : String(e)}`,
-    };
-  }
-
-  return { ok: true, warning: switchWarning };
 }
 
 /** Import a mind from a .volute archive (extracted to tempDir by CLI). */
@@ -2267,35 +2030,13 @@ const app = new Hono<AuthEnv>()
     if (!isKnownTemplate(template)) {
       return c.json({ error: `Unknown template: ${template}` }, 400);
     }
-    const UPGRADE_BRANCH = "upgrade";
-    const upgradeVariantName = `${mindName}-upgrade`;
-    const worktreeDir = resolve(dir, ".variants", UPGRADE_BRANCH);
 
     if (body.abort) {
-      if (!existsSync(worktreeDir)) {
+      if (!upgradeInProgress(mindName)) {
         return c.json({ error: "No upgrade in progress" }, 400);
       }
-
       try {
-        // Abort merge if mid-merge
-        try {
-          const gitDirContent = readFileSync(resolve(worktreeDir, ".git"), "utf-8").trim();
-          const gitDir = gitDirContent.replace("gitdir: ", "");
-          if (existsSync(resolve(gitDir, "MERGE_HEAD"))) {
-            await gitExec(["merge", "--abort"], { cwd: worktreeDir });
-          }
-        } catch {}
-
-        await cleanupVariant(upgradeVariantName, mindName, dir, worktreeDir, { stop: true });
-
-        // Also delete the upgrade branch directly — cleanupVariant uses the variant
-        // name as fallback branch, but the actual branch is UPGRADE_BRANCH
-        try {
-          await gitExec(["branch", "-D", UPGRADE_BRANCH], { cwd: dir });
-        } catch {
-          // Branch may already be deleted by cleanupVariant
-        }
-
+        await abortUpgrade(mindName);
         return c.json({ ok: true });
       } catch (err) {
         return c.json(
@@ -2306,105 +2047,48 @@ const app = new Hono<AuthEnv>()
     }
 
     if (body.continue) {
-      // Continue upgrade after conflict resolution — merge back to main
-      if (!existsSync(worktreeDir)) {
+      if (!upgradeInProgress(mindName)) {
         return c.json({ error: "No upgrade in progress" }, 400);
       }
-
-      const status = await gitExec(["status", "--porcelain"], { cwd: worktreeDir });
-      const hasConflicts = status
-        .split("\n")
-        .some((line) => line.startsWith("UU") || line.startsWith("AA"));
-      if (hasConflicts) {
-        return c.json({ error: "Unresolved conflicts remain" }, 409);
-      }
-
       try {
-        await gitExec(["add", "-A"], { cwd: worktreeDir });
-        await gitExec(["commit", "-m", "merge template update"], { cwd: worktreeDir });
-      } catch (e) {
-        const msg = e instanceof Error ? e.message : String(e);
-        const stderr = (e as any)?.stderr ?? "";
-        const stdout = (e as any)?.stdout ?? "";
-        if (
-          !msg.includes("nothing to commit") &&
-          !stderr.includes("nothing to commit") &&
-          !stdout.includes("nothing to commit")
-        )
-          throw e;
-      }
-
-      // Re-add home files that match the new .gitignore allowlist patterns
-      try {
-        await gitExec(["add", "home/"], { cwd: worktreeDir });
+        const result = await continueUpgrade(mindName, { template });
+        if (result.status === "conflicts") {
+          return c.json({
+            ok: false,
+            conflicts: true,
+            worktreeDir: result.worktreeDir,
+            message: "Merge conflicts detected. Resolve them, then run with continue.",
+          });
+        }
+        return c.json({ ok: true, warning: result.warning });
       } catch (err) {
-        log.warn(`failed to re-add home files during upgrade for ${mindName}`, log.errorData(err));
-      }
-      try {
-        await gitExec(["diff", "--cached", "--quiet"], { cwd: worktreeDir });
-      } catch {
-        await gitExec(["commit", "-m", "re-add allowlisted home files"], {
-          cwd: worktreeDir,
-        });
-      }
-
-      // Fix ownership after root git operations
-      await chownMindDir(dir, mindName);
-
-      // Merge upgrade branch back to main, cleanup, and restart
-      try {
-        const result = await mergeUpgradeAndRestart(
-          mindName,
-          dir,
-          worktreeDir,
-          upgradeVariantName,
-          UPGRADE_BRANCH,
-          template,
-          oldTemplate,
-        );
-        return c.json(result);
-      } catch (err) {
-        return c.json(
-          { error: err instanceof Error ? err.message : "Failed to merge upgrade" },
-          500,
-        );
+        const msg = err instanceof Error ? err.message : "Failed to merge upgrade";
+        if (msg === "Unresolved conflicts remain") {
+          return c.json({ error: msg }, 409);
+        }
+        return c.json({ error: msg }, 500);
       }
     }
 
     if (body.accept) {
       // Legacy — upgrades now auto-merge. Clean up any old-style upgrade state.
-      if (existsSync(worktreeDir)) {
+      if (upgradeInProgress(mindName)) {
         try {
-          await cleanupVariant(upgradeVariantName, mindName, dir, worktreeDir, { stop: true });
+          await abortUpgrade(mindName);
         } catch (err) {
           log.warn(`failed to clean up legacy upgrade variant for ${mindName}`, log.errorData(err));
         }
-        try {
-          await gitExec(["branch", "-D", UPGRADE_BRANCH], { cwd: dir });
-        } catch {}
       }
       return c.json({ error: "Upgrades now auto-merge. Run 'volute mind upgrade' again." }, 400);
     }
 
     if (body.diff) {
-      // Preview what the upgrade would change
+      // Initialize git repo if missing
+      if (!existsSync(resolve(dir, ".git"))) {
+        return c.json({ error: "Mind has no git history — nothing to diff against" }, 400);
+      }
       try {
-        // Initialize git repo if missing
-        if (!existsSync(resolve(dir, ".git"))) {
-          return c.json({ error: "Mind has no git history — nothing to diff against" }, 400);
-        }
-
-        await updateTemplateBranch(dir, template, mindName);
-
-        // Show what the template branch has that main doesn't
-        let diff: string;
-        try {
-          diff = await gitExec(["diff", "HEAD...volute/template"], { cwd: dir });
-        } catch {
-          // If three-dot diff fails (no common ancestor), fall back to two-dot
-          diff = await gitExec(["diff", "HEAD", "volute/template"], { cwd: dir });
-        }
-
+        const diff = await upgradeDiff(mindName, template);
         return c.json({ ok: true, diff: diff || "(no changes)" });
       } catch (err) {
         return c.json(
@@ -2415,132 +2099,25 @@ const app = new Hono<AuthEnv>()
     }
 
     // Fresh upgrade
-
-    if (existsSync(worktreeDir)) {
+    if (upgradeInProgress(mindName)) {
       return c.json(
         { error: "Upgrade variant already exists. Use continue or delete it first." },
         409,
       );
     }
 
-    // Initialize git repo if missing (minds created before git config was fixed)
-    if (!existsSync(resolve(dir, ".git"))) {
-      try {
-        const env = isIsolationEnabled()
-          ? { ...process.env, HOME: resolve(dir, "home") }
-          : undefined;
-        await gitExec(["init"], { cwd: dir, mindName: mindName, env });
-        await configureGitIdentity(mindName, { cwd: dir, mindName: mindName, env });
-        await gitExec(["add", "-A"], { cwd: dir, mindName: mindName, env });
-        await gitExec(["commit", "-m", "initial commit"], { cwd: dir, mindName: mindName, env });
-        await chownMindDir(dir, mindName);
-      } catch (err) {
-        rmSync(resolve(dir, ".git"), { recursive: true, force: true });
-        return c.json(
-          {
-            error: `Git initialization failed: ${err instanceof Error ? err.message : String(err)}`,
-          },
-          500,
-        );
-      }
-    }
-
-    // Clean up stale worktree refs and leftover branch
-    await gitExec(["worktree", "prune"], { cwd: dir });
     try {
-      await gitExec(["branch", "-D", UPGRADE_BRANCH], { cwd: dir });
-    } catch {
-      // branch doesn't exist
-    }
-
-    // Update template branch
-    await updateTemplateBranch(dir, template, mindName);
-
-    // Create upgrade worktree
-    const parentDir = resolve(dir, ".variants");
-    if (!existsSync(parentDir)) {
-      mkdirSync(parentDir, { recursive: true });
-    }
-
-    await gitExec(["worktree", "add", "-b", UPGRADE_BRANCH, worktreeDir], { cwd: dir });
-
-    // Prepare home/ allowlist migration: untrack home files so template
-    // branch removal doesn't cause conflicts or deletions
-    await gitExec(["rm", "-r", "--cached", "--ignore-unmatch", "home/"], {
-      cwd: worktreeDir,
-    });
-    // Re-add VOLUTE.md so template merge can update it
-    try {
-      await gitExec(["checkout", "HEAD", "--", "home/VOLUTE.md"], { cwd: worktreeDir });
-      await gitExec(["add", "home/VOLUTE.md"], { cwd: worktreeDir });
-    } catch (err) {
-      const msg = String((err as Error)?.message ?? err);
-      if (!msg.includes("did not match")) {
-        log.warn(
-          `unexpected error restoring VOLUTE.md during upgrade for ${mindName}`,
-          log.errorData(err),
-        );
-      }
-    }
-    // Commit prep step if there are changes
-    try {
-      await gitExec(["diff", "--cached", "--quiet"], { cwd: worktreeDir });
-    } catch {
-      await gitExec(["commit", "-m", "prepare for home/ allowlist migration"], {
-        cwd: worktreeDir,
-      });
-    }
-
-    // Merge template branch
-    const hasConflicts = await mergeTemplateBranch(worktreeDir);
-
-    if (!hasConflicts) {
-      // Re-add home files that match the new .gitignore allowlist patterns
-      try {
-        await gitExec(["add", "home/"], { cwd: worktreeDir });
-      } catch (err) {
-        log.warn(`failed to re-add home files during upgrade for ${mindName}`, log.errorData(err));
-      }
-      try {
-        await gitExec(["diff", "--cached", "--quiet"], { cwd: worktreeDir });
-      } catch {
-        await gitExec(["commit", "-m", "re-add allowlisted home files"], {
-          cwd: worktreeDir,
+      const result = await runUpgrade(mindName, { template });
+      if (result.status === "conflicts") {
+        return c.json({
+          ok: false,
+          conflicts: true,
+          worktreeDir: result.worktreeDir,
+          message: "Merge conflicts detected. Resolve them, then run with continue.",
         });
       }
-    }
-
-    // Fix ownership — daemon runs as root but mind needs to own its files
-    await chownMindDir(dir, mindName);
-
-    if (hasConflicts) {
-      return c.json({
-        ok: false,
-        conflicts: true,
-        worktreeDir,
-        message: "Merge conflicts detected. Resolve them, then run with continue.",
-      });
-    }
-
-    // Merge upgrade branch back to main, cleanup, and restart
-    try {
-      const result = await mergeUpgradeAndRestart(
-        mindName,
-        dir,
-        worktreeDir,
-        upgradeVariantName,
-        UPGRADE_BRANCH,
-        template,
-        oldTemplate,
-      );
-      return c.json(result);
+      return c.json({ ok: true, warning: result.warning });
     } catch (err) {
-      // Merge failed — clean up
-      try {
-        await cleanupVariant(upgradeVariantName, mindName, dir, worktreeDir);
-      } catch (cleanupErr) {
-        log.warn(`cleanup failed after upgrade error for ${mindName}`, log.errorData(cleanupErr));
-      }
       return c.json({ error: err instanceof Error ? err.message : "Failed to merge upgrade" }, 500);
     }
   })
